A job to fix up the index.
When we deployed the new v9 schema and series index, it accidentally wrote new format index entries to the old index rows, causing query errors.

This PR introduces a job to delete them.

Also add DeleteBatch and QueryRows to the StorageClient interface and HashKey to the query results; only implemented on BigTable.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>